check in the following leak fix -- it never got any feedback on the lists but I've been running with it for months now...
--------------- ORIGINAL EMAIL:

Through the magic of printf-debugging I tracked my leak down to
some buffers remaining in the AL_PROCESSED state even after being
unqueued and deleted, because their buf->current_list.items==1.
So they get flagged as ALB_PENDING_DELETE but that final
unref never comes...

This patch seems to completely plug my leak and doesn't introduce
any (new) strangeness (to my ear or valgrind).

I'm a little concerned about it though (not least because of
the comment it replaces/contradicts ;)).  My main concern is that
the RemoveCurrentRef() might just be wallpapering over the
real bug, for example that the mixer (the only place where AL ever
AddCurrentRef()s) isn't (finalizing or) Removing the ref properly
(though it looks like it should) when a queue runs dry.

However, since a buffer can't be both AL_PROCESSED and 'current'
at the same time on the same source (right?) this workaround
at unqueue-time should be harmless (removing a ref that doesn't
exist works out as a no-op).  Or for all I know this patch is
completely the right thing to do... the sample implementation
is still a bit of a tangled mystery to me.

_alSourceQueueClear() might need the same 'fix', if this patch
is acceptable.

Feedback welcome (jv!).  I can check it in if it looks okay.


git-svn-id: svn://connect.creativelabs.com/trunk@340 109fce75-3012-0410-ac6a-932efec96c1b