Synchronized the type of psnd_pcm_hw_params_get_period_size with the corresponding ALSA 1.x prototype.
ALSA versioning is still handled incorrectly when OPENAL_DLOPEN_ALSA
is defined: Consider the case when e.g. the library was compiled
against a 1.x ALSA (i.e. SND_LIB_MAJOR != 0), but at runtime dlvsym()
finds a ALSA_0.9.0rc4 version of a given function. The function is
then still called like it has a 1.x signature, leading to incorrect
parameter passing.

The right way to handle this versionitis would be to remember which
version was found and use the corresponding interface later. If only
an unversioned symbol could be found, assuming the 1.x signature is
probably the right thing nowadays.


git-svn-id: svn://connect.creativelabs.com/trunk@411 109fce75-3012-0410-ac6a-932efec96c1b