A bunch of API cleanups:
Changed the first parameter of alutInit and alutInitWithoutContext to
int*, not ALint*. argc is an int, after all.

Changed return type of alutGetError to ALenum, this is more consistent
with alGetError.

Changed the argument of alutGetErrorString to ALenum for the same
reason.

Changed the first parameter of alutLoadMemoryFromFileImage and
alutCreateBufferFromFileImage to "const ALvoid*", this is more
convenient and more consistent with alBufferData.

Changed the first parameter of alutLoadWAVFile and alutLoadWAVMemory
back to ALbyte*, like it was before.

Changed the frequency parameter of alutLoadMemoryFromFile and
alutLoadMemoryFromFileImage to an ALuint*, this is more consistent
with the way AL handles frequencies.

Changed the return type of alutLoadMemoryFromFile and
alutLoadMemoryFromFileImage to an ALvoid* for more consistency.


git-svn-id: svn://connect.creativelabs.com/trunk@576 109fce75-3012-0410-ac6a-932efec96c1b